backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
neo4j==5.23.0
numpy==2.1.2
pydantic==2.9.2
